+++ R/get_pull_requests.R
@@ -15,7 +15,7 @@
 #'     token = Sys.getenv("GITLAB_PAT_PUBLIC"),
 #'     orgs = "mbtests"
 #'   )
-#'  get_issues(my_gitstats, since = "2018-01-01", state = "open")
+#'  get_pull_requests(my_gitstats, since = "2018-01-01", state = "open")
 #' }
 #' @export
 get_pull_requests <- function(gitstats,