+++ NEWS.md
@@ -6,6 +6,7 @@
 - Added `get_pull_requests_stats()` function ([#726](https://github.com/r-world-devs/GitStats/issues/726)).
 - Reorganized fixtures and test helpers ([#727](https://github.com/r-world-devs/GitStats/issues/727)).
 - Prettified messages with new icons ([#148](https://github.com/r-world-devs/GitStats/issues/148), [#361](https://github.com/r-world-devs/GitStats/issues/361)).
+- Refactored code for getting repositories with code to make it more readable ([#612]https://github.com/r-world-devs/GitStats/issues/612)).
 
 # GitStats 2.3.9